# docs/api/openapi/cordum-api.yaml
openapi: 3.0.3
info:
  title: Cordum HTTP API
  version: '2026-04-21.2'
  description: Canonical OpenAPI 3.0.3 spec for the Cordum gateway HTTP surface.
  contact:
    name: Cordum Team
  license:
    name: Proprietary
servers:
- url: http://localhost:8081
  description: Gateway HTTP (local default)
- url: /
  description: Relative gateway base
...
security:
- apiKey: []
- bearerAuth: []
paths:
...
  /api/v1/chat:
    post:
      operationId: postLLMChat
      summary: Send a single LLM chat message
      description: Proxies an authenticated single-shot chat turn to cordum-llm-chat.
      tags:
      - LLMChat
      parameters:
      - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - message
              properties:
                session_id:
                  type: string
                message:
                  type: string
      responses:
        '200':
          description: Chat turn result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TierLimit'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/chat/stream:
    get:
      operationId: streamLLMChat
      summary: Stream an LLM chat turn with Server-Sent Events
      tags:
      - LLMChat
      parameters:
      - $ref: '#/components/parameters/TenantID'
      - name: session_id
        in: query
        schema:
          type: string
      - name: message
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: SSE stream of chat frames
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TierLimit'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/chat/ws:
    get:
      operationId: websocketLLMChat
      summary: Open the LLM chat assistant WebSocket
      description: Upgrades to a WebSocket and streams chat frames for the Gmail-style dashboard widget.
      tags:
      - LLMChat
      parameters:
      - $ref: '#/components/parameters/TenantID'
      - name: session_id
        in: query
        schema:
          type: string
      responses:
        '101':
          description: WebSocket upgrade
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TierLimit'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/chat/sessions:
    get:
      operationId: listLLMChatSessions
      summary: List LLM chat sessions
      tags:
      - LLMChat
      parameters:
      - $ref: '#/components/parameters/TenantID'
      - name: cursor
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          default: 50
      responses:
        '200':
          description: Chat session page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key issued by the Cordum admin.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login.
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant isolation header (required on all protected routes).
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request — invalid or missing parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized — missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden — insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    BadGateway:
      description: Bad gateway — upstream service error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceUnavailable:
      description: Service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TierLimit:
      description: Forbidden by current license entitlements
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericObject'
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        status:
          type: integer
          description: HTTP status code
      required:
      - error
      - status
...
    GenericObject:
      type: object
      properties: {}
      additionalProperties: true
...